Add spike removal to emboss plane
This takes the spike removal that was in the command line script and
adds it to the main plugin.  In order to see the effects of the displacement
modifier a "copy" of the mesh data with everything applied is used.

This is a bit on the slow side, at the moment.  I might be able to speed
it up by only looking at the "emboss" vertex group.