Update name plate creation
This updated name plate code is also fully integrated into emboss plane
so the plate and model can be made at the same time.

The external edge code has been replaced as a special case of a name
plate with no text.

The holder code was also updated to run in 2.8X