Fix #16 and parent objects
This PR fixes #16 by searching for the image texture with the same base
name as the model rather than assuming it will be first in the texture
list.

Also, all external objects are now parented so the base model can be
moved around the scene and all the objects will be moved with it (e.g.
name plate, frame, text).  This also means that the base rotation of the
plane can be preserved when making models.